+++ packages/core/src/types/match.ts
@@ -1,0 +1,19 @@
+import { InternalRoute } from "./route";
+import { Params, MissResponse, PendingMatchResponse } from "./response";
+
+export interface MatchingRoute {
+  route: InternalRoute;
+  params: Params;
+}
+
+export interface MissMatch {
+  route: undefined;
+  response: MissResponse;
+}
+
+export interface BestMatch {
+  route: InternalRoute;
+  response: PendingMatchResponse;
+}
+
+export type Match = MissMatch | BestMatch;