+++ packages/core/tests/curi.spec.ts
@@ -168,7 +168,7 @@ describe("curi", () => {
 
       describe("sideEffects", () => {
         it("calls side effect methods AFTER a response is generated, passing them response and navigation", done => {
-          const routes = [{ name: "All", path: ":all+" }];
+          const routes = [{ name: "All", path: "(.*)" }];
           const sideEffect = jest.fn();
 
           const router = curi(history, routes, {
@@ -182,8 +182,8 @@ describe("curi", () => {
           });
         });
 
-        it('calls side effects WITHOUT "after: true" property before response handlers', done => {
-          const routes = [{ name: "All", path: ":all+" }];
+        it('calls side effects WITHOUT "after: true" BEFORE response handlers', done => {
+          const routes = [{ name: "All", path: "(.*)" }];
 
           const sideEffect1 = jest.fn();
           const sideEffect2 = jest.fn();
@@ -203,11 +203,11 @@ describe("curi", () => {
           });
         });
 
-        it('calls side effects WITH "after: true" property AFTER response handlers', done => {
+        it('calls side effects WITH "after: true" AFTER response handlers', done => {
           const routes = [
             {
               name: "All",
-              path: ":all+",
+              path: "(.*)",
               on: {
                 initial: () => Promise.resolve() // force async
               }
@@ -226,7 +226,7 @@ describe("curi", () => {
         });
 
         it("passes response, navigation, and router object to side effect", done => {
-          const routes = [{ name: "All", path: ":all*" }];
+          const routes = [{ name: "All", path: "(.*)" }];
           const responseHandler = jest.fn();
           const sideEffect = function({ response, navigation, router }) {
             expect(response).toMatchObject({
@@ -592,7 +592,7 @@ describe("curi", () => {
     });
 
     it("passes response, navigation, and router object to response handler", done => {
-      const routes = [{ name: "All", path: ":all*" }];
+      const routes = [{ name: "All", path: "(.*)" }];
       const responseHandler = function({ response, navigation, router }) {
         expect(response).toMatchObject({
           name: "All",