+++ packages/core/types/utils/match.d.ts
@@ -1,3 +1,4 @@
 import { HickoryLocation } from "@hickory/root";
-import { InternalRoute, BestMatch } from "../types/route";
-export default function matchLocation(location: HickoryLocation, routes: Array<InternalRoute>): BestMatch;
+import { InternalRoute } from "../types/route";
+import { Match } from "../types/match";
+export default function matchLocation(location: HickoryLocation, routes: Array<InternalRoute>): Match;