+++ packages/core/types/utils/parseParams.d.ts
@@ -1,3 +1,3 @@
-import { ParamParsers } from '../types/route';
-import { RawParams, Params } from '../types/response';
-export default function parseParams(params: RawParams, fns: ParamParsers): Params;
+import { ParamParsers } from "../types/route";
+import { RawParams, Params } from "../types/response";
+export default function parseParams(params: RawParams, fns?: ParamParsers): Params;