+++ packages/core/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* [TypeScript] Add generic type to describe shape of `response.body` to help enforce that each response has the same `body` format.
 * Move `match.response` to top level and group `match.every` and `match.initial` under `on` object (`on.initial` and `on.every`).
 * Group `initial` and `every` resolved values with `error` under `resolved` object.
 * Rename `router.refresh` to `router.replaceRoutes`.